requirements.txt
requests
Pillow
pydantic
python-dotenv
rich